separated pickup code into another script called PickUp
