+++ ICS 161 Game 3/Assets/Prefabs/Player.prefab
@@ -47,7 +47,22 @@ GameObject:
   - component: {fileID: 114970459288258118}
   m_Layer: 0
   m_Name: Camera
-  m_TagString: Untagged
+  m_TagString: camera
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!1 &1585034185682038
+GameObject:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  serializedVersion: 5
+  m_Component:
+  - component: {fileID: 4633545636647438}
+  m_Layer: 0
+  m_Name: Hold Slot
+  m_TagString: hold slot
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -61,7 +76,8 @@ Transform:
   m_LocalRotation: {x: 0.06289886, y: -0, z: -0, w: 0.99801993}
   m_LocalPosition: {x: 0, y: 0.7, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
+  m_Children:
+  - {fileID: 4633545636647438}
   m_Father: {fileID: 4588055076219664}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 7.2120004, y: 0, z: 0}
@@ -71,12 +87,25 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1169542933764228}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 1, z: 0}
+  m_LocalRotation: {x: -0, y: -0.99937785, z: -0, w: -0.035270628}
+  m_LocalPosition: {x: 36.39, y: 1, z: 27.78}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 4246596787385080}
   m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: -184.043, z: 0}
+--- !u!4 &4633545636647438
+Transform:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1585034185682038}
+  m_LocalRotation: {x: -0.06289886, y: -0, z: -0, w: 0.99801993}
+  m_LocalPosition: {x: 1, y: -0.29651916, z: 0.76326686}
+  m_LocalScale: {x: 0.5000007, y: 0.5, z: 0.5000007}
+  m_Children: []
+  m_Father: {fileID: 4246596787385080}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!20 &20772512447186140
@@ -97,7 +126,7 @@ Camera:
     height: 1
   near clip plane: 0.3
   far clip plane: 1000
-  field of view: 60
+  field of view: 70
   orthographic: 0
   orthographic size: 5
   m_Depth: -1
