+++ ICS 161 Game 3/Assets/Prefabs/Weapon.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bf69bfeb439af104db5280aec0b5523a
+timeCreated: 1489024991
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
