+++ ICS 161 Game 3/Assets/Scripts/PickUp.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 690d8c54b2b83744fb51b7ee89e31bf1
+timeCreated: 1489025260
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
